Plugins: categorize linux.check_creds as a malwarep lugin
